requirements.txt
appdirs==1.4.3
attrs==18.2.0
black==18.6b4
click==6.7
Flask==1.0.2
Flask-GoogleRecaptcha==0.5.8
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
pluggy==0.7.1
py==1.6.0
python-dotenv==0.9.1
six==1.11.0
toml==0.9.6
tox==3.4.0
virtualenv==16.0.0
Werkzeug==0.14.1
